Examples for Genre Classification and Revenue Regression
Added an example for genre classication, both as binary classification
(comedy or not-comedy) and multi-class (predicting the primary genre
from a selection of several genres).

Also added gross revenue for a subset of movies from IMDb in
`data/actual_gross.csv`, which we use to build a simple linear
regression model for actors -> revenue.